+++ react/components/Logo.js
@@ -1,38 +1,38 @@
 import React from 'react'
-import PropTypes from 'prop-types'
 import { ExtensionPoint, Link } from 'vtex.render-runtime'
-import { CONSTANTS } from './Helpers'
-import useDevice from '../hooks/useDevice'
+import PropTypes from 'prop-types'
+import { logo } from '../defaults'
 
-import header from '../store-header.css'
+import styles from '../store-header.css'
 
-const Logo = ({ link, src, title, size }) => {
-  const { mobile, desktop } = useDevice()
-
+/**
+ * Represents the header logo
+ */
+const Logo = ({ logoUrl, logoTitle, linkUrl, logoSize, mobile }) => {
   return (
-    <div className={`${header.topMenuLogo} pv2 mr5`}>
-      <Link to={link} className={`outline-0 ${header.logoLink}`}>
+    <div className={`${styles.topMenuLogo} pv2 mr5`}>
+      <Link to={linkUrl} className={`outline-0 ${styles.logoLink}`}>
         {mobile && (
           <div className="db dn-ns">
             <ExtensionPoint
               id="logo"
-              url={src}
-              title={title}
-              width={size.mobile.width}
-              height={size.mobile.height}
+              url={logoUrl}
+              title={logoTitle}
+              width={logoSize.mobile.width}
+              height={logoSize.mobile.height}
               isMobile={true}
             />
           </div>
         )}
 
-        {desktop && (
+        {!mobile && (
           <div className="dn db-ns">
             <ExtensionPoint
               id="logo"
-              url={src}
-              title={title}
-              width={size.desktop.width}
-              height={size.desktop.height}
+              url={logoUrl}
+              title={logoTitle}
+              width={logoSize.desktop.width}
+              height={logoSize.desktop.height}
               isMobile={false}
             />
           </div>
@@ -43,34 +43,13 @@ const Logo = ({ link, src, title, size }) => {
 }
 
 Logo.propTypes = {
-  src: PropTypes.string.isRequired,
-  link: PropTypes.string,
-  title: PropTypes.string,
-  size: PropTypes.shape({
-    desktop: PropTypes.shape({
-      width: PropTypes.number,
-      height: PropTypes.number,
-    }),
-    mobile: PropTypes.shape({
-      width: PropTypes.number,
-      height: PropTypes.number,
-    }),
-  }),
+  /** If it's mobile mode */
+  mobile: PropTypes.bool.isRequired,
+  ...logo.propTypes,
 }
 
 Logo.defaultProps = {
-  link: '/',
-  title: '',
-  size: {
-    desktop: {
-      width: CONSTANTS.LOGO.DESKTOP.WIDTH,
-      height: CONSTANTS.LOGO.DESKTOP.HEIGHT,
-    },
-    mobile: {
-      width: CONSTANTS.LOGO.MOBILE.WIDTH,
-      height: CONSTANTS.LOGO.MOBILE.HEIGHT,
-    },
-  },
+  ...logo.defaultProps,
 }
 
 export default Logo