+++ react/hooks/useDevice.js
@@ -1,25 +1,24 @@
 import { useState, useEffect } from 'react'
-import { useRuntime } from 'vtex.render-runtime'
-import { CONSTANTS } from '../components/Helpers'
+import { debounce } from 'debounce'
 
 /**
  * Hook that handles dynamic device change on resize
- * @param {*} mobileBreakpoint : Break point of mobile change
+ * @param {Number} mobileBreakpoint - break point of mobile change
+ * @returns {Boolean, Boolean} - if its mobile or desktop
  */
-const useDevice = (mobileBreakpoint = CONSTANTS.RESIZE_BREAKPOINTS.MOBILE) => {
-  const { hints } = useRuntime()
-  const [mobile, setMobile] = useState(hints.mobile)
-  const [desktop, setDesktop] = useState(hints.desktop)
+const useDevice = (mobileBreakpoint = 640) => {
+  const [mobile, setMobile] = useState(window.innerWidth <= mobileBreakpoint)
+  const [desktop, setDesktop] = useState(window.innerWidth > mobileBreakpoint)
 
   useEffect(() => {
     const handleResize = () => {
       setMobile(window.innerWidth <= mobileBreakpoint)
       setDesktop(window.innerWidth > mobileBreakpoint)
     }
-
-    window.addEventListener('resize', handleResize)
+    const debounced = debounce(handleResize, 100)
+    window.addEventListener('resize', debounced)
     return () => {
-      window.removeEventListener('resize', handleResize)
+      window.removeEventListener('resize', debounced)
     }
   }, [mobileBreakpoint])